Automated g4 rollback of changelist 316462109.
*** Reason for rollback ***

This CL includes a proto that is compiled with the open source compiler that does not understand szl_api_version.  I believe I need to do something like cl/316423607 for at least one of the proto files involved.  For now, let's roll back in order to not inconvenience the team.

*** Original change description ***

Explicitly set szl_api_version = 2.

There is still some Sawzall code running in production as a business logic execution engine in GMail and Gaia.  We need to manually finish the proto2 conversion in order to release a new Sawzall binary that is compatible with new changes to protocol buffers.

More information:  ​go/sawzall-proto2-lsc and go/sawzall-proto2

Tested:
    tap_presubmit: http://test/OCL:315701415:BASE:316210669:1592040780158:d61d033f
    Some tests failed; test failures are beli...

***

PiperOrigin-RevId: 316652609